requirements.txt
Django==1.10
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.5.1
billiard==3.3.0.23
celery==3.1.23
django-crispy-forms==1.6.0
django-registration-redux==1.4
djangorestframework==3.4.4
gunicorn==19.6.0
kombu==3.0.35
psycopg2==2.6.2
pytz==2016.6.1
requests==2.11.0
ua-parser==0.7.1
user-agents==1.0.1
wsgiref==0.1.2
